+++ embedded/xsens_imu/example/test.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
-#include "imu.h"
+#include "../lib/imu.h"
 
 int main() {
 	SetupIMU();
